requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
pdf2image
pillow
flask
